// src/models/hour.ts
export class Hour {
   
    isChecked: boolean; //0
    memberName: string;
    ssn: string;
    kHours: number;
    kAmount: number;
    flex: number;  //5
    fullLocal: string;
    workDate: Date;
    kRate: number;
    flexRate: number;
    entryDate: Date; //10
    empAccountNum: string;
    empId: number;
    mhrsId: number;
    mbrId: number;
}